pyproject.toml
[project]
name = "model-pipelines"
version = "0.1.0"
description = ""
authors = [
    {name = "Ashwin Senthilvasan",email = "ashwin.senthilvasan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio (>=1.71.0,<2.0.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "torch (>=1.8.0)",
    "realesrgan @ git+https://github.com/asenthilvasan/Real-ESRGAN.git",
    "pandas (>=1.1.4)",
    "requests (>=2.32.2)",
    "pillow (>=10.3.0)",
    "numpy (>=1.23.5)",
    "opencv-python-headless (>=4.11.0.86,<5.0.0.0)",
    "gitpython (>=3.1.30)",
    "matplotlib (>=3.3)",
    "opencv-python (>=4.1.1)",
    "pyyaml (>=5.3.1)",
    "scipy (>=1.4.1)",
    "thop (>=0.1.1)",
    "torchvision (>=0.9.0)",
    "tqdm (>=4.66.3)",
    "ultralytics (>=8.2.34)",
    "seaborn (>=0.11.0)",
    "setuptools (>=70.0.0)",
    "psutil",
    "aiohttp (>=3.12.9,<4.0.0)"
]

[tool.poetry]
packages = [{include = "model_pipelines", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
